add "watch" button and stub out events
there's a known issue with filesystem event watchers in
general where they can erroneously send multiple events
due to other watchers interacting with them. for the
moment i'm not going to attempt to "debounce" these
events, and simply eat the cost of poking the database ~3
times when a file is created.

another thing i noticed when implementing is that
"rename" events on directories don't cascade events into
files/subfolders in that directory, so i had to manually
enumerate through them. hopefully this should suffice
for my uses, at least initially. if it's a problem i can revisit
it later.