(busted) trigger database updates from filesystem events
this currently doesn't work. on my system, when a file is
copied into the watch directory, three events are triggered:
one creation, and two change events. these three events
race with each other to set database state, and also happen
in different threads, causing null-reference exceptions due
to incomplete database connection information. (i think?
i'm not super sure.)

FileSystemWatcher can synchronize all its events to a UI
thread via its SynchronizationObject, but WPF forms
don't implement the required interface to be used as a
synchronization object. this commit attempts to dispatch
events to the UI thread via the form's Dispatcher, but this
just caused the Npgsql exceptions to be hidden from the
debugger instead. future work will introduce a dedicated
dispatcher, to make sure that event threads can access
a properly-initialized Database object, as well as attempt
to debounce events. it will still need to marshal signals
back to MainWindow to update state later on when
queries are added, but that should be doable.