do database update on the UI thread + add cancellation
by making the FileSystemWatcher's events just queue their
actions into a queue that's then read by the button handler,
the database events all see the proper connection state. this
also makes it easier to make them update the UI later on,
since they'll be acting in the UI thread.

this change also turns the "Watch Dir" button into a cancel
button while the FileSystemWatcher is active. if the user
clicks the button again while a watcher is listening, the
form will clear the watcher state and cancel the inner
event loop listening on the TaskQueue.