add ReplayGain tags to track info
since TagLib# doesn't allow accessing "custom tags"
through the unified interface, these have to be read via the
format-specific mechanisms. thankfully all the metadata
is stored in the format-specific tag containers, so with a
little foreknowledge of the tag format and the fields we're
looking for, we can load in the ones we want.